Better fast-forward in timer_cx_read for CX II
While rescheduling immediately after timer reads speeds up some loads,
on the CX II it causes a massive slowdown. There is a timer read in various
IO functions to handle timeouts, so this triggers a translation exit
and reschedule even though the code immediately following realized that it's
already done. A compromise is to shorten the time until the next reschedule
on each timer read, which doesn't immediately trigger in the timeout checks.